fix: SQLDelight INTEGER columns are Long, not Int (CI fix #5)
SQLDelight generates INTEGER columns as Long (not Int). LibraryCategory's
displayOrder field was typed as Int, but the generated row access returns Long.
Changed displayOrder to Long in both LibraryCategory + LibraryItem data classes.
Also simplified addToDefaultCategory to pass the Long directly.

+++ ANI-KUTA/APP/ani-kuta/core/content/src/main/java/com/confused/anikuta/core/content/ContentModels.kt
@@ -112,7 +112,7 @@ data class ExtensionInfo(
 data class LibraryCategory(
     val id: Long,
     val name: String,
-    val displayOrder: Int,
+    val displayOrder: Long,
     val isPermanent: Boolean,
     val createdAt: Long,
 )
@@ -124,6 +124,6 @@ data class LibraryItem(
     val id: Long,
     val mainId: String,
     val categoryId: Long,
-    val displayOrder: Int,
+    val displayOrder: Long,
     val addedAt: Long,
 )
